修复部分剧集无法加载
fixed #149

+++ bili-api/src/main/kotlin/dev/aaa1115910/biliapi/http/entity/season/AppSeasonData.kt
@@ -366,7 +366,7 @@ data class AppSeasonData(
         val play: String,
         val reply: Int,
         val share: Int,
-        val views: Int,
+        val views: Long,
         val vt: Int
     )
